HADOOP-13897. s3-cse
Contributed by Igor Mazur.

This is HADOOP-13897-trunk-013.patch merged in to trunk so that conflicts
are resolved. It does not compile; that wasn't its goal.

I've done this so that I can start a new branch which pulls in some changes
(docs, config options, tests) but otherwise handles S3-CSE very differently

* unpadded algorithm only. This vastly simplifes the patch as
  there's no need to handle any code in production or test where file length !=
  data written; --0 bytes still means directory.
* CSE options become something we can marshall in DTs, so integrate there.

all we should need to do is
-load options and secrets
-set them when creating every request
-new tests