HADOOP-19317. S3A: add option to enable/disable 100 CONTINUE
Option fs.s3a.connection.expect.continue controls
whether or not an HTTP PUT request to the S3 store
Sets the Expect: 100-continue
header and awaits a 100 CONTINUE response
from the server before uploading any data.

This allows for throttling and other problems
to be detected fast.

The default is "true".
This has long been the implicit setting;
this change simply allows callers to disable it.